test: move aws cni unit tests to its own package
